Move conditional-edge dedup tests into builder_test.go
